Update the page URL when selecting user
When you select a user and press the back button of your browser, it
will exit out of schedule page instead of going back inside of the
app. This is unexpected behavior.

This commit will update the URL using history.pushState() when a user
is selected. Allowing the browser navigation buttons to nicely
integrate with the schedule.